feat: 规整目录结构
