+++ lib/pages/Chatbot/chatbot_page.dart
@@ -1,12 +1,12 @@
 // 这个文件定义了chatbot的核心功能部分，包括聊天页面和不同对话方式的相关逻辑。
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:namer_app/global_setting.dart';
+import 'package:namer_app/constants/global_setting.dart';
 import 'package:namer_app/models/chat_models.dart';
 import 'package:namer_app/providers/progress.dart';
 import 'chat_widgets.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
-import '../providers/user_preference.dart';
+import '../../providers/user_preference.dart';
 
 // ChatbotPage class是chatbot的主页面，负责显示聊天界面和处理用户输入。
 class ChatbotPage extends ConsumerStatefulWidget {
