+++ lib/pages/DietMonitoring/binge_eating_card.dart
