+++ lib/pages/DietMonitoring/binge_eating_options.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:namer_app/DietMonitoring/diet_monitoring_page.dart';
-import 'package:namer_app/DietMonitoring/food_purge.dart';
+import 'package:namer_app/pages/DietMonitoring/binge_eating_card.dart';
+import 'package:namer_app/pages/DietMonitoring/binge_eating_record_page.dart';
 
-class MonitoringOptionsPage extends StatelessWidget {
+class BingeEatingOptions extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('监控记录', style: TextStyle(color: Colors.black)),
+        title: Text('冲动记录', style: TextStyle(color: Colors.black)),
         elevation: 0,
       ),
       body: Center(
@@ -20,15 +20,15 @@ class MonitoringOptionsPage extends StatelessWidget {
                 Navigator.push(
                     context,
                     MaterialPageRoute(
-                        builder: (context) => DietMonitoringPage()));
-              }, // 增加字体大小
+                        builder: (context) => BingeEatingRecordPage()));
+              },
               style: ElevatedButton.styleFrom(
                 foregroundColor: Colors.white,
                 backgroundColor: Color.fromARGB(255, 174, 86, 165), // 按钮文字颜色
                 padding: EdgeInsets.symmetric(
                     horizontal: 32, vertical: 16), // 增加按钮的padding来增大按钮
               ),
-              child: Text('监控记录', style: TextStyle(fontSize: 20)),
+              child: Text('冲动记录', style: TextStyle(fontSize: 20)), // 增加字体大小
             ),
             SizedBox(height: 20),
             ElevatedButton(
@@ -38,15 +38,15 @@ class MonitoringOptionsPage extends StatelessWidget {
                     context,
                     MaterialPageRoute(
                         builder: (context) =>
-                            FoodPurgePage())); // 假设FoodPurgePage是食物清除页面的Widget
-              }, // 增加字体大小
+                            BingeEatingCard())); // 假设FoodPurgePage是食物清除页面的Widget
+              },
               style: ElevatedButton.styleFrom(
                 foregroundColor: Colors.white,
                 backgroundColor: Color.fromARGB(255, 217, 156, 211), // 按钮文字颜色
                 padding: EdgeInsets.symmetric(
                     horizontal: 32, vertical: 16), // 增加按钮的padding来增大按钮
               ),
-              child: Text('食物清除', style: TextStyle(fontSize: 20)),
+              child: Text('冲动应对卡片', style: TextStyle(fontSize: 20)), // 增加字体大小
             ),
           ],
         ),
